requirements.txt
fastmcp
mcp
google-generativeai
python-dotenv
